webapp/requirements.txt
django
djangorestframework
mongoengine
requests